AI Risk

+++ Application/DTOs/AI/Risk/RiskCategoryDto.cs
@@ -3,9 +3,9 @@
     public class RiskCategoryDto
     {
         public string Category { get; set; } = string.Empty;
-        public int Score { get; set; }
+        public double Score { get; set; }   
         public string Severity { get; set; } = string.Empty;
         public List<string> ContributingFactors { get; set; } = new();
-        public int Weight { get; set; }
+        public double Weight { get; set; }
     }
 }
