+++ Application/DTOs/AI/Risk/RiskScoresDto.cs
@@ -2,9 +2,9 @@
 {
     public class RiskScoresDto
     {
-        public int Overall { get; set; }
+        public double Overall { get; set; } 
         public string Severity { get; set; } = string.Empty;
         public List<RiskCategoryDto> Categories { get; set; } = new();
-        public int Confidence { get; set; }
+        public double Confidence { get; set; }
     }
 }
