+++ Application/DTOs/AI/Risk/SimilarCaseDto.cs
@@ -3,7 +3,7 @@
     public class SimilarCaseDto
     {
         public string ProjectName { get; set; } = string.Empty;
-        public int SimilarityScore { get; set; }
+        public double SimilarityScore { get; set; }           
         public string Outcome { get; set; } = string.Empty;
         public string KeyLesson { get; set; } = string.Empty;
     }
